ENH: allow planning pyFFTW builders with unitary normalisation
