+++ pyfftw/builders/builders.py
@@ -256,7 +256,8 @@ The exceptions raised by each of these functions are as per their
 equivalents in :mod:`numpy.fft`, or as documented above.
 '''
 
-from ._utils import _precook_1d_args, _Xfftn
+from ._utils import _precook_1d_args, _Xfftn, _norm_args
+
 
 __all__ = ['fft','ifft', 'fft2', 'ifft2', 'fftn',
            'ifftn', 'rfft', 'irfft', 'rfft2', 'irfft2', 'rfftn',
@@ -266,7 +267,7 @@ __all__ = ['fft','ifft', 'fft2', 'ifft2', 'fftn',
 def fft(a, n=None, axis=-1, overwrite_input=False,
         planner_effort='FFTW_MEASURE', threads=1,
         auto_align_input=True, auto_contiguous=True,
-        avoid_copy=False):
+        avoid_copy=False, norm=None):
     '''Return a :class:`pyfftw.FFTW` object representing a 1D FFT.
 
     The first three arguments are as per :func:`numpy.fft.fft`;
@@ -280,12 +281,12 @@ def fft(a, n=None, axis=-1, overwrite_input=False,
 
     return _Xfftn(a, s, axes, overwrite_input, planner_effort,
             threads, auto_align_input, auto_contiguous,
-            avoid_copy, inverse, real)
+            avoid_copy, inverse, real, **_norm_args(norm))
 
 def ifft(a, n=None, axis=-1, overwrite_input=False,
         planner_effort='FFTW_MEASURE', threads=1,
         auto_align_input=True, auto_contiguous=True,
-        avoid_copy=False):
+        avoid_copy=False, norm=None):
     '''Return a :class:`pyfftw.FFTW` object representing a 1D
     inverse FFT.
 
@@ -301,13 +302,13 @@ def ifft(a, n=None, axis=-1, overwrite_input=False,
 
     return _Xfftn(a, s, axes, overwrite_input, planner_effort,
             threads, auto_align_input, auto_contiguous,
-            avoid_copy, inverse, real)
+            avoid_copy, inverse, real, **_norm_args(norm))
 
 
 def fft2(a, s=None, axes=(-2,-1), overwrite_input=False,
         planner_effort='FFTW_MEASURE', threads=1,
         auto_align_input=True, auto_contiguous=True,
-        avoid_copy=False):
+        avoid_copy=False, norm=None):
     '''Return a :class:`pyfftw.FFTW` object representing a 2D FFT.
 
     The first three arguments are as per :func:`numpy.fft.fft2`;
@@ -321,12 +322,12 @@ def fft2(a, s=None, axes=(-2,-1), overwrite_input=False,
 
     return _Xfftn(a, s, axes, overwrite_input, planner_effort,
             threads, auto_align_input, auto_contiguous,
-            avoid_copy, inverse, real)
+            avoid_copy, inverse, real, **_norm_args(norm))
 
 def ifft2(a, s=None, axes=(-2,-1), overwrite_input=False,
         planner_effort='FFTW_MEASURE', threads=1,
         auto_align_input=True, auto_contiguous=True,
-        avoid_copy=False):
+        avoid_copy=False, norm=None):
     '''Return a :class:`pyfftw.FFTW` object representing a
     2D inverse FFT.
 
@@ -341,13 +342,13 @@ def ifft2(a, s=None, axes=(-2,-1), overwrite_input=False,
 
     return _Xfftn(a, s, axes, overwrite_input, planner_effort,
             threads, auto_align_input, auto_contiguous,
-            avoid_copy, inverse, real)
+            avoid_copy, inverse, real, **_norm_args(norm))
 
 
 def fftn(a, s=None, axes=None, overwrite_input=False,
         planner_effort='FFTW_MEASURE', threads=1,
         auto_align_input=True, auto_contiguous=True,
-        avoid_copy=False):
+        avoid_copy=False, norm=None):
     '''Return a :class:`pyfftw.FFTW` object representing a n-D FFT.
 
     The first three arguments are as per :func:`numpy.fft.fftn`;
@@ -361,12 +362,12 @@ def fftn(a, s=None, axes=None, overwrite_input=False,
 
     return _Xfftn(a, s, axes, overwrite_input, planner_effort,
             threads, auto_align_input, auto_contiguous,
-            avoid_copy, inverse, real)
+            avoid_copy, inverse, real, **_norm_args(norm))
 
 def ifftn(a, s=None, axes=None, overwrite_input=False,
         planner_effort='FFTW_MEASURE', threads=1,
         auto_align_input=True, auto_contiguous=True,
-        avoid_copy=False):
+        avoid_copy=False, norm=None):
     '''Return a :class:`pyfftw.FFTW` object representing an n-D
     inverse FFT.
 
@@ -381,12 +382,12 @@ def ifftn(a, s=None, axes=None, overwrite_input=False,
 
     return _Xfftn(a, s, axes, overwrite_input, planner_effort,
             threads, auto_align_input, auto_contiguous,
-            avoid_copy, inverse, real)
+            avoid_copy, inverse, real, **_norm_args(norm))
 
 def rfft(a, n=None, axis=-1, overwrite_input=False,
         planner_effort='FFTW_MEASURE', threads=1,
         auto_align_input=True, auto_contiguous=True,
-        avoid_copy=False):
+        avoid_copy=False, norm=None):
     '''Return a :class:`pyfftw.FFTW` object representing a 1D
     real FFT.
 
@@ -403,12 +404,12 @@ def rfft(a, n=None, axis=-1, overwrite_input=False,
 
     return _Xfftn(a, s, axes, overwrite_input, planner_effort,
             threads, auto_align_input, auto_contiguous,
-            avoid_copy, inverse, real)
+            avoid_copy, inverse, real, **_norm_args(norm))
 
 def irfft(a, n=None, axis=-1, overwrite_input=False,
         planner_effort='FFTW_MEASURE', threads=1,
         auto_align_input=True, auto_contiguous=True,
-        avoid_copy=False):
+        avoid_copy=False, norm=None):
     '''Return a :class:`pyfftw.FFTW` object representing a 1D
     real inverse FFT.
 
@@ -425,12 +426,12 @@ def irfft(a, n=None, axis=-1, overwrite_input=False,
 
     return _Xfftn(a, s, axes, overwrite_input, planner_effort,
             threads, auto_align_input, auto_contiguous,
-            avoid_copy, inverse, real)
+            avoid_copy, inverse, real, **_norm_args(norm))
 
 def rfft2(a, s=None, axes=(-2,-1), overwrite_input=False,
         planner_effort='FFTW_MEASURE', threads=1,
         auto_align_input=True, auto_contiguous=True,
-        avoid_copy=False):
+        avoid_copy=False, norm=None):
     '''Return a :class:`pyfftw.FFTW` object representing a 2D
     real FFT.
 
@@ -444,12 +445,12 @@ def rfft2(a, s=None, axes=(-2,-1), overwrite_input=False,
 
     return _Xfftn(a, s, axes, overwrite_input, planner_effort,
             threads, auto_align_input, auto_contiguous,
-            avoid_copy, inverse, real)
+            avoid_copy, inverse, real, **_norm_args(norm))
 
 def irfft2(a, s=None, axes=(-2,-1),
         planner_effort='FFTW_MEASURE', threads=1,
         auto_align_input=True, auto_contiguous=True,
-        avoid_copy=False):
+        avoid_copy=False, norm=None):
     '''Return a :class:`pyfftw.FFTW` object representing a 2D
     real inverse FFT.
 
@@ -466,13 +467,13 @@ def irfft2(a, s=None, axes=(-2,-1),
 
     return _Xfftn(a, s, axes, overwrite_input, planner_effort,
             threads, auto_align_input, auto_contiguous,
-            avoid_copy, inverse, real)
+            avoid_copy, inverse, real, **_norm_args(norm))
 
 
 def rfftn(a, s=None, axes=None, overwrite_input=False,
         planner_effort='FFTW_MEASURE', threads=1,
         auto_align_input=True, auto_contiguous=True,
-        avoid_copy=False):
+        avoid_copy=False, norm=None):
     '''Return a :class:`pyfftw.FFTW` object representing an n-D
     real FFT.
 
@@ -487,13 +488,13 @@ def rfftn(a, s=None, axes=None, overwrite_input=False,
 
     return _Xfftn(a, s, axes, overwrite_input, planner_effort,
             threads, auto_align_input, auto_contiguous,
-            avoid_copy, inverse, real)
+            avoid_copy, inverse, real, **_norm_args(norm))
 
 
 def irfftn(a, s=None, axes=None,
         planner_effort='FFTW_MEASURE', threads=1,
         auto_align_input=True, auto_contiguous=True,
-        avoid_copy=False):
+        avoid_copy=False, norm=None):
     '''Return a :class:`pyfftw.FFTW` object representing an n-D
     real inverse FFT.
 
@@ -510,4 +511,4 @@ def irfftn(a, s=None, axes=None,
 
     return _Xfftn(a, s, axes, overwrite_input, planner_effort,
             threads, auto_align_input, auto_contiguous,
-            avoid_copy, inverse, real)
+            avoid_copy, inverse, real, **_norm_args(norm))
