+++ pyfftw/interfaces/_utils.py
@@ -46,6 +46,7 @@ import pyfftw
 import numpy
 from . import cache
 
+
 def _Xfftn(a, s, axes, overwrite_input, planner_effort,
         threads, auto_align_input, auto_contiguous,
         calling_func, normalise_idft=True, ortho=False):
